requirements.txt
numpy<2.0
opencv-python

easyocr
pillow
tensorflow
scikit-learn

selenium
matplotlib
seaborn